+++ kedro/contrib/idempotent/idempotent_runner.py
@@ -40,37 +40,40 @@ from kedro.pipeline.node import Node
 from kedro.runner.runner import AbstractRunner
 
 
-def run_node_idempotently(node: Node, catalog: DataCatalog, state: IdempotentStateStorage) -> Node:
+def run_node_idempotently(node: Node, catalog: DataCatalog, state: IdempotentStateStorage, force_run: bool) -> Node:
     """Run a single `Node` with inputs from and outputs to the `catalog`.
 
     Args:
         node: The ``Node`` to run.
         catalog: A ``DataCatalog`` containing the node's inputs and outputs.
         state: An ``IdempotentStateStorage`` to reference node runs
+        force_run: boolean to force run all node regardless of state
 
     Returns:
         The node argument.
 
     """
-    # Find all inputs that are parameters
-    parameter_inputs = [i for i in node.inputs if i.startswith("params:")]
-    # Hash them as the key after loading
-    for parameter_input in parameter_inputs:
-        # Update our idempotency state with new hashes
-        state.update_run_id(parameter_input, str(hash(catalog.load(parameter_input))))
 
-    # Find all output that are MemoryDataSet
-    memory_outputs = [o for o in node.outputs if type(catalog._data_sets[o]) == MemoryDataSet]
+    if not force_run:
+        # Find all inputs that are parameters
+        parameter_inputs = [i for i in node.inputs if i.startswith("params:")]
+        # Hash them as the key after loading
+        for parameter_input in parameter_inputs:
+            # Update our idempotency state with new hashes
+            state.update_run_id(parameter_input, str(hash(catalog.load(parameter_input))))
 
-    # Judge if the node should be run
-    inputs_have_changed = state.node_inputs_have_changed(
-        node.name,
-        node.inputs
-    )
-    has_been_run = state.node_has_been_run(node.name)
-    should_run_node = not has_been_run or inputs_have_changed or len(memory_outputs) > 0
-    if not should_run_node:
-        return node
+        # Find all output that are MemoryDataSet
+        memory_outputs = [o for o in node.outputs if type(catalog._data_sets[o]) == MemoryDataSet]
+
+        # Judge if the node should be run
+        inputs_have_changed = state.node_inputs_have_changed(
+            node.name,
+            node.inputs
+        )
+        has_been_run = state.node_has_been_run(node.name)
+        should_run_node = not has_been_run or inputs_have_changed or len(memory_outputs) > 0
+        if not should_run_node:
+            return node
 
     inputs = {name: catalog.load(name) for name in node.inputs}
     outputs = node.run(inputs)
@@ -96,6 +99,7 @@ class IdempotentSequentialRunner(AbstractRunner):
 
     def __init__(self):
         super().__init__()
+        self.force_run = False
         self.state_storage = IdempotentStateStorage()
 
     def create_default_data_set(self, ds_name: str) -> AbstractDataSet:
@@ -128,7 +132,7 @@ class IdempotentSequentialRunner(AbstractRunner):
 
         for exec_index, node in enumerate(nodes):
             try:
-                run_node_idempotently(node, catalog, self.state_storage)
+                run_node_idempotently(node, catalog, self.state_storage, self.force_run)
                 done_nodes.add(node)
             except Exception:
                 self._suggest_resume_scenario(pipeline, done_nodes)
@@ -146,3 +150,9 @@ class IdempotentSequentialRunner(AbstractRunner):
             self._logger.info(
                 "Completed %d out of %d tasks", exec_index + 1, len(nodes)
             )
+
+
+class ForcedIdempotentSequentialRunner(IdempotentSequentialRunner):
+    def __init__(self):
+        super().__init__()
+        self.force_run = True
